refactor(config): move ConventionResolver from services to config
Moves ConventionResolver to config package where it belongs, as it only
depends on config modules (ProfileConvention, ProfileConfig), clients,
adapters, and exceptions - all lower-layer packages.

Changes:
- Move convention_resolver.py from services/ to config/
- Create backward-compat re-export shim at old location
- Update services/__init__.py lazy import mapping
- Update config/__init__.py to export ConventionResolver
- Update agent prompt imports to use new location

Backward compatibility preserved via re-export shim, so existing imports
from vibe3.services.convention_resolver continue to work.

Issue: #1885
