+++ src/vibe3/config/__init__.py
@@ -1,5 +1,7 @@
 """Config package."""
 
+from typing import Any
+
 from vibe3.config.loader import get_config, load_config, reload_config
 from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.config.settings import (
@@ -30,4 +32,31 @@ __all__ = [
     "QualityConfig",
     "PRScoringConfig",
     "MergeGateConfig",
+    # Lazy imports
+    "ConventionResolver",
 ]
+
+# Lazy import mapping for symbols to avoid circular dependencies
+_SYMBOL_MODULES = {
+    "ConventionResolver": "vibe3.config.convention_resolver",
+}
+
+
+def __getattr__(name: str) -> Any:
+    """Lazy import for config symbols to avoid circular dependencies.
+
+    This allows external modules to use:
+        from vibe3.config import ConventionResolver
+
+    While avoiding circular imports at module load time.
+    """
+    if name in _SYMBOL_MODULES:
+        import importlib
+
+        module = importlib.import_module(_SYMBOL_MODULES[name])
+        symbol = getattr(module, name)
+        # Cache in module globals for faster subsequent access
+        globals()[name] = symbol
+        return symbol
+
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
