+++ tests/vibe3/config/test_policy_paths.py
@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 
+from vibe3.config.convention_resolver import ConventionResolver
 from vibe3.config.settings import PlanConfig, ReviewConfig, RunConfig
-from vibe3.services.convention_resolver import ConventionResolver
 
 
 def test_convention_resolver_vibe_center_policy_path() -> None:
