+++ tests/vibe3/config/test_profile_config.py
@@ -1,8 +1,8 @@
 """Tests for profile-based resource resolution."""
 
 from vibe3.adapters import get_adapter
+from vibe3.config.convention_resolver import ConventionResolver
 from vibe3.config.profile_config import ProfileConfig
-from vibe3.services.convention_resolver import ConventionResolver
 
 
 def test_profile_config_vibe_center_loads_adapter() -> None:
